Convert useCreateTitanMailboxMutation to Typescript
